Add Authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -64,6 +64,14 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        // if (\Gate::allows('update-question', $question)) {
+        //     return view('questions.edit', compact('question'));
+        // }
+        // return abort(403, 'Access denied');
+
+        if (\Gate::denies('update-question', $question)) {
+            return abort(403, 'Access denied');
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -89,6 +97,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            return abort(403, 'Access denied');
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Question deleted successfully');
     }
